Move files containing component definitions to a components directory
