+++ main_window/components/__init__.py
@@ -1,0 +1,2 @@
+from .labels import *
+from .plot_info import *
