+++ main_window/components/labels.py
@@ -12,7 +12,7 @@ from abc import ABC, abstractmethod
 from PySide6.QtWidgets import QLabel, QLayout
 from PySide6.QtCore import Qt
 
-from .utils import packet_spec
+from ..utils import packet_spec
 
 class DynamicLabel(ABC):
     # Acts as a wrapper class for QLabel components that gives them
