+++ main_window/components/plot_info.py
